Frontier is now a COM client.   - call initCOM() before using any COM interfaces   - a thread global keeps track of whether or not     COM has already been initialized for the current     thread (so it won't happen more than once)   - disposeglobals closes down the COM connection
The table of character sets at string.charsets is now
for informational purposes only. The item names are
the internet names, and the value is a 'display name'
for the character set.

The string.convertCharset verb now takes internet names
instead of code page ID's.

New verb string.isCharsetAvailabe(). Give it an internet
name, it returns a boolean.

All changes for Windows only. Two platforms are mostly
in sync now, but for a couple tweaks still lacking
on the mac.

git-svn-id: https://frontierkernel.svn.sourceforge.net/svnroot/frontierkernel/Frontier/trunk@1328 4e289971-d10f-0410-980f-25908192c0f7