update version to 10.1b17
So, I noticed that all the version properties for the different Apps
that could be built on top of Frontier were all the same, so I
consolidated them at the top of the file.  I also bumped the build
status from alpha to beta.  The version out in the wild for some time
has been the alpha build, so it seems time to bump to beta.

At any rate, I should be incrementing the version number every time I
post a build, so that practice can start now.

Signed-off-by: Ted C. Howard <ted@tedchoward.com>

git-svn-id: https://frontierkernel.svn.sourceforge.net/svnroot/frontierkernel/Frontier/trunk@1839 4e289971-d10f-0410-980f-25908192c0f7