Removed the applet and iowa runtimes
Build the Mac version the same way as the windows version. This could
introduce some breakage in usertalk scripts that depend on the 'card'
verbs being present for the Mac platform.

This is a step towards simplification. The less code is in the project,
the less I have to understand.