Load string resources from strings file
We are no longer relying on reading string values from STR# resources in
the resource fork. These strings have been copied to Localizable.strings
and the code now reads from the strings file. The strings are keyed
based on the STR# resource id and the index of the string in the list.