The About Window is not a Cocoa window
Yes, I went a little crazy, but this turned out to not be as intrusive
as I initially thought. The rest of the app still functions the same.
The about window still exposes the same C api, it's just implemented in
Objective-C on the Mac now. It's actually pretty cool.

It's also not complete. The popup isn't implemented, but with the way
scheduler2 constantly sends ccmsg calls, it really isn't useful anyway.
There are also some C functions that currently don't do anything. Also,
the zoom functionality is not implemented. You just get the full-size
about window with icon and copyright info along with the ccmsg.