Rewrote folderloop() to not use FSSpec
This was fairly involved because the old way of looping through a folder
depended on setting a file index on a folder's FSSpec ParamBlock. The
new way involves using an FSIterator.

I am nervous about this commit because I don't know how to test it. I
tried several things in Frontier, but I never managed to execute this
block. I wonder if I spent an afternoon rewriting code that never gets
called.