Rewrote largefilebuffer()
This function depended on the Mac's FreeMem() and CompactMem()
functions. These functions became no-ops once we transitioned to Mac OSX
and virtual memory. In the modern world (ignore the iPhone for the
moment) we can safely allocate as much memory as we need without worry.

The function now always allocates 128K.