Support Building on Windows 10 (#11)
Frontier wasn't building on Windows 10. Apparently there were some 
changes to the C runtime provided by the platform, and some
requirements around byte alignment to link to the Windows API.

Using the VC2K5 Solution as a base (since it includes the OPML Editor
targets) I created a new Visual Studio 2022 solution. After making the
necessary adjustments to the project configuration to be able to use
the new C runtime and link to the Windows SDK, I had to wrap the struct
definitions in #pragma pack(2) so that the bytes were aligned properly
to load .root files from disk.

There were also some issues with loading DLL files. Apparently the
Windows version of the OPML editor ships with several binary DLL files.
I found references to two of them in the source code.

- `COMDll.dll` seems to be needed for COM support
- `webappdll.dll` seems to be used to render the accordion player in
the background of the main window

For now, I have disabled this functionality. The app builds and runs.